Replaced autonomy gain in Cultural Sufferance with min autonomy
This was done since increasing the autonomy right after the colony is completed also set the "hidden" autonomy to that number. In other words colonies start with 50 autonomy if they are in a state and it drops to 0 once you core them. With the privilege they instead got 70% autonomy which didn't drop to 20% once cored, thus the effective increase used to be 70% not 20%.